CURB-5590 minor changes after review feedback
