+++ extensions/@shopgate-product-reviews/frontend/package.json
@@ -15,6 +15,6 @@
     "react-dom": "^17.0.2"
   },
   "dependencies": {
-    "lodash": "^4.17.21"
+    "lodash": "^4.17.23"
   }
 }
