+++ extensions/@shopgate-theme-config/frontend/package.json
@@ -4,7 +4,7 @@
   "description": "Provides the theme configurations for Catalog 2.0",
   "license": "Apache-2.0",
   "devDependencies": {
-    "lodash": "^4.17.21"
+    "lodash": "^4.17.23"
   },
   "peerDependencies": {
     "@shopgate/pwa-common": ">=7.20.0"
