+++ extensions/@shopgate-user-privacy/frontend/package.json
@@ -14,7 +14,7 @@
     "prop-types": "~15.8.1",
     "react": "^17.0.2",
     "react-redux": "^8.1.3",
-    "redux": "^4.2.1"
+    "redux": "^5.0.1"
   },
   "version": "7.30.0"
 }
