+++ libraries/common/package.json
@@ -29,7 +29,7 @@
     "path-match": "^1.2.4",
     "react-dotdotdot": "~1.3.1",
     "react-helmet": "^6.1.0",
-    "react-portal": "^4.3.0",
+    "react-portal": "^3.1.0",
     "react-redux": "^8.1.3",
     "redux": "^5.0.1",
     "redux-logger": "^3.0.6",
