+++ libraries/core/package.json
@@ -16,7 +16,7 @@
   ],
   "dependencies": {
     "detector": "^2.5.0",
-    "lodash": "^4.17.21",
+    "lodash": "^4.17.23",
     "mobile-detect": "^1.4.3"
   }
 }
