+++ libraries/engage/checkout/components/ShippingMethods/ShippingMethods.jsx
@@ -4,7 +4,6 @@ import React, {
 import PropTypes from 'prop-types';
 import { css } from 'glamor';
 import classNames from 'classnames';
-import { hashString } from '@shopgate/pwa-common/helpers/hash';
 import sortBy from 'lodash/sortBy';
 import uniqBy from 'lodash/uniqBy';
 import { themeConfig } from '@shopgate/pwa-common/helpers/config';
@@ -13,6 +12,7 @@ import {
   RadioGroupV2 as RadioGroup, RadioCard, MessageBar,
 } from '@shopgate/engage/components';
 import { useCheckoutContext } from '@shopgate/engage/checkout/hooks/common';
+import { hashString } from '@shopgate/pwa-core/helpers';
 import ShippingMethod from './ShippingMethod';
 import connect from './connector';
