+++ libraries/engage/product/selectors/variants.js
@@ -1,7 +1,17 @@
-import { createSelector } from 'reselect';
+import { createSelector, setGlobalDevModeChecks } from 'reselect';
 import find from 'lodash/find';
 import { getProductVariants } from '@shopgate/pwa-common-commerce/product';
 import isEqual from 'lodash/isEqual';
+
+// In Reselect v5, dev-mode checks call input selectors multiple times.
+// In unit tests (where input selectors are often mocked with `mockReturnValueOnce`),
+// this can cause noisy warnings/failures unrelated to the memoization behavior we want to test.
+if (process.env.NODE_ENV === 'test') {
+  setGlobalDevModeChecks({
+    inputStabilityCheck: 'never',
+    identityFunctionCheck: 'never',
+  });
+}
 
 /**
  * Creates a selector that retrieves a product by a characteristic.
