+++ themes/theme-ios11/package.json
@@ -48,7 +48,7 @@
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-helmet": "^6.1.0",
-    "react-portal": "^4.3.0",
+    "react-portal": "^3.1.0",
     "react-redux": "^8.1.3",
     "react-transition-group": "^3.0.0",
     "redux": "^5.0.1",
