+++ themes/theme-ios11/pages/Favorites/components/EmptyFavorites/components/Icon/index.jsx
@@ -1,5 +1,5 @@
 import React, { useMemo } from 'react';
-import { hashString } from '@shopgate/pwa-common/helpers/hash';
+import { hashString } from '@shopgate/pwa-core/helpers';
 import styles from './style';
 
 /**
